microservices/requirements.txt
google-cloud-discoveryengine==0.13.9
Jinja2==3.1.6
google-adk==1.5.0